converted int fields to decimal

+++ LitigationSoftware/LS.Models/ITReturnDetails.cs
@@ -6,7 +6,6 @@ namespace LS.Models
     public class ITReturnDetails : BaseEntity
     {
         #region Properties
-
         public int CompanyID { get; set; }
         public string CompanyName { get; set; }
         public string PANNumber { get; set; }
@@ -17,28 +16,28 @@ namespace LS.Models
         public string ITSectionDescription { get; set; }
         public DateTime? ITReturnFillingDate { get; set; }
         public DateTime? ITReturnDueDate { get; set; }
-        public int? HousePropIncome { get; set; }
-        public int? IncomefromCapGainsNonSTT { get; set; }
-        public int? IncomefromCapGainsSTT { get; set; }
+        public decimal? HousePropIncome { get; set; }
+        public decimal? IncomefromCapGainsNonSTT { get; set; }
+        public decimal? IncomefromCapGainsSTT { get; set; }
         public bool? IncomefromBusinessProf { get; set; }
-        public int? UnabsorbedDepreciation { get; set; }
-        public int? Broughtforwardlosses { get; set; }
-        public int? IncomeFromOtherSources { get; set; }
-        public int? DeductChapterVIA { get; set; }
-        public int? ProfitUS115JB { get; set; }
-        public int? AdvanceTax1installment { get; set; }
-        public int? AdvanceTax2installment { get; set; }
-        public int? AdvanceTax3installment { get; set; }
-        public int? AdvanceTax4installment { get; set; }
-        public int? TDS { get; set; }
-        public int? TCSPaidbyCompany { get; set; }
-        public int? SelfAssessmentTax { get; set; }
-        public int? MATCredit { get; set; }
-        public int? InterestUS234A { get; set; }
-        public int? InterestUS234B { get; set; }
-        public int? InterestUS234C { get; set; }
-        public int? InterestUS244A { get; set; }
-        public int? RefundReceived { get; set; }
+        public decimal? UnabsorbedDepreciation { get; set; }
+        public decimal? Broughtforwardlosses { get; set; }
+        public decimal? IncomeFromOtherSources { get; set; }
+        public decimal? DeductChapterVIA { get; set; }
+        public decimal? ProfitUS115JB { get; set; }
+        public decimal? AdvanceTax1installment { get; set; }
+        public decimal? AdvanceTax2installment { get; set; }
+        public decimal? AdvanceTax3installment { get; set; }
+        public decimal? AdvanceTax4installment { get; set; }
+        public decimal? TDS { get; set; }
+        public decimal? TCSPaidbyCompany { get; set; }
+        public decimal? SelfAssessmentTax { get; set; }
+        public decimal? MATCredit { get; set; }
+        public decimal? InterestUS234A { get; set; }
+        public decimal? InterestUS234B { get; set; }
+        public decimal? InterestUS234C { get; set; }
+        public decimal? InterestUS244A { get; set; }
+        public decimal? RefundReceived { get; set; }
         public bool? RevisedReturnFile { get; set; }
         public bool IsReturn { get; set; }
